requirements.txt
matplotlib==3.8.0
numpy==1.26.1
pandas==2.1.1
Pillow==10.1.0
scikit-learn==1.3.1
tensorboard==2.15.1
# torch==2.1.0
# torchaudio==2.1.0
# torchvision==0.16.0
tqdm==4.66.1
